requirements.txt
flask
python-dotenv
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
langchain-core
faiss-cpu
pypdf
sentence-transformers
